detection/requirements.txt
tensorflow==1.12.0
pillow
protobuf
Cython
contextlib2
lxml
jupyter
matplotlib
pyyaml